count to decimal

+++ app/models/spree/assign_part_to_bundle_form.rb
@@ -52,7 +52,7 @@ module Spree
     end
 
     def quantity
-      part_options[:count].to_i
+      part_options[:count].to_d
     end
 
     def assemblies_part
